+++ language/paypal_english.php
@@ -28,6 +28,5 @@ $LANG_SHOP_gateway = array(
 'prod_url'      => 'PayPal Url - Production',
 'sandbox_url'   => 'PayPal Url - Sandbox',
 'test_mode'     => 'Testing (Sandbox) Mode?',
-'ipn_url'       => 'IPN URL Override',
 );
 ?>